MealCandidateResultsおよびTripCandidateResultsコンポーネントを追加し、ResultPageでの候補別結果表示を改善
